final commit

+++ controllers/contact.js
@@ -3,15 +3,15 @@ import Contact from "#models/contact";
 
 // @desc Get ALL Contacts
 // @route GET /contact/get-contacts
-// @access PUBLIC
+// @access PRIVATE
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 
 // @desc Get single contact
 // @route GET /contact/get-contact/:id
-// @access PUBLIC
+// @access PRIVATE
 const getContact = asyncHandler(async (req, res) => {
   const id = req.params.id;
 
@@ -22,12 +22,17 @@ const getContact = asyncHandler(async (req, res) => {
     throw new Error("Contact not found");
   }
 
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to view other user contacts");
+  }
+
   res.status(200).json(contact);
 });
 
 // @desc Add new contact
 // @route POST /contact/add-contact
-// @access PUBLIC
+// @access PRIVATE
 const addContact = asyncHandler(async (req, res) => {
   const { name, email, phone } = req.body;
   if (!name || !email || !phone) {
@@ -39,6 +44,7 @@ const addContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    user_id: req.user.id,
   });
 
   res
@@ -48,7 +54,7 @@ const addContact = asyncHandler(async (req, res) => {
 
 // @desc Update Contact
 // @route PUT /contact/update-contact/:id
-// @access PUBLIC
+// @access PRIVATE
 const updateContact = asyncHandler(async (req, res) => {
   const id = req.params.id;
 
@@ -65,7 +71,14 @@ const updateContact = asyncHandler(async (req, res) => {
     throw new Error("All fields are mandotary");
   }
 
-  const newContact = await Contact.findByIdAndUpdate(req.params.id, req.body);
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to update other user contacts");
+  }
+
+  const newContact = await Contact.findByIdAndUpdate(req.params.id, req.body, {
+    new: true,
+  });
 
   res
     .status(200)
@@ -74,7 +87,7 @@ const updateContact = asyncHandler(async (req, res) => {
 
 // @desc Delete Contact
 // @route DELETE /contact/delete-contact/:id
-// @access PUBLIC
+// @access PRIVATE
 const deleteContact = asyncHandler(async (req, res) => {
   const id = req.params.id;
 
@@ -83,6 +96,11 @@ const deleteContact = asyncHandler(async (req, res) => {
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
+  }
+
+  if (contact.user_id.toString() !== req.user.id) {
+    res.status(403);
+    throw new Error("User don't have permission to delete other user contacts");
   }
 
   await Contact.deleteOne({ _id: id });
